Move collision detection utilities from hooks/ to utils/
The functions checkHorizontalOverlap and checkAABBCollision are pure
functions with no React hook semantics. Move them to utils/collision.ts
to better reflect their nature and improve code organization.

Update all import paths in use-bombs.ts and use-items.ts accordingly,
and update README.md to reflect the new file location.

Closes #131
